AX: WebProcess::accessibilityRelayProcessSuspended can drop requests to unsuspend, causing ATs to think web content is permanently suspended https://bugs.webkit.org/show_bug.cgi?id=268024 rdar://problem/121545631
Reviewed by Andres Gonzalez.

Depending on timing, we can get a call to unsuspend the process at a moment when we don't have
any webpages (but may gain them soon). Even when we do get a webpage back, we don't try to send
the process unsuspension message, causing assistive technologies to think we are permanently suspended.

Fix this by storing this pending unsuspension as a new WebProcess member variable, and processing it
when a webpage is added.

* Source/WebKit/WebProcess/WebProcess.cpp:
(WebKit::WebProcess::createWebPage):
(WebKit::WebProcess::accessibilityRelayProcessSuspended):
* Source/WebKit/WebProcess/WebProcess.h:

Canonical link: https://commits.webkit.org/273474@main