365scores.com: Quirks needed to silence the resize events when putting Safari in the background (iPadOS) https://bugs.webkit.org/show_bug.cgi?id=268130 rdar://116491386
Reviewed by Brent Fulgham.

When setting the 365scores.com in the background/foreground on iPad,
some resize events are being fired. These events in return modifies
the layout when it should stay the same. This patch add this domain
to the existing quirk to solve this type of issues.

* Source/WebCore/page/Quirks.cpp:
(WebCore::Quirks::shouldSilenceWindowResizeEvents const):

Canonical link: https://commits.webkit.org/273567@main