Fix JSCOnly build on Windows https://bugs.webkit.org/show_bug.cgi?id=268189
Reviewed by Yusuke Suzuki.

Add `win/Win32Handle.cpp` to the source listing for Windows builds.

Fix the guard of `RunLoop::registerRunLoopMessageWindowClass()` which
is `USE(WINDOWS_EVENT_LOOP)` not `OS(WINDOWS)`.

Don't override the default for `ENABLE_WEBASSEMBLY` because it is not
available on x86. It will default to `ON` for x64.

* Source/WTF/wtf/PlatformJSCOnly.cmake:
* Source/WTF/wtf/Threading.cpp:
* Source/cmake/OptionsJSCOnly.cmake:

Canonical link: https://commits.webkit.org/273585@main