Add some extra hardening to prevent crashes in -[WKScrollingNodeScrollViewDelegate axesToPreventScrollingForPanGestureInScrollView:] https://bugs.webkit.org/show_bug.cgi?id=268229 rdar://121329521
Reviewed by Aditya Keerthi.

While re-reading some of the crash logs in rdar://121329521, I realized that the speculative
mitigation in 273443@main only takes care of the case where the scroll view is either directly
removed with a call to `-removeFromSuperview`, or no longer in the view hierarchy. However, it's
possible that a view above the `WKBaseScrollView` was unparented, which might cause UIKit to
internally transition the axis locking pan gesture to either `.failed` or `.cancelled` state and
invoke the action.

Add some extra hardening to be robust in this scenario as well, by bailing in the case where the
axis-locking pan gesture is either cancelled or failed.

* Source/WebKit/UIProcess/ios/WKBaseScrollView.mm:
(-[WKBaseScrollView _updatePanGestureToPreventScrolling]):

Canonical link: https://commits.webkit.org/273613@main