IPC validation for DocumentEditingContext::Range should not crash on overflow https://bugs.webkit.org/show_bug.cgi?id=268300 rdar://121859057
Reviewed by Chris Dumez.

It's currently possible to crash the UI process by having it attempt to decode a
`DocumentEditingContext::Range` that overflows the `uint64_t` limit. This is because the validator
currently creates a `Checked<uint64_t>`, which (by default) uses the `CrashOnOverflow` handler. This
should use `RecordOverflow` instead, which we can get by adopting the `CheckedUint64` type alias.

* Source/WebKit/Shared/DocumentEditingContext.serialization.in:

Canonical link: https://commits.webkit.org/273671@main