Crash under VideoPresentationManagerProxy::createLayerHostViewWithID() https://bugs.webkit.org/show_bug.cgi?id=268320 rdar://121747495
Reviewed by Per Arne Vollan.

Null check the GPUProcess before dereferencing it. Even though we had a
GPUProcess to generate the id in the first place, the ID comes from IPC
from the WebProcess and things are just racing. We shouldn't assume the
GPUProcess is still running by the time we get the IPC.

* Source/WebKit/UIProcess/Cocoa/VideoPresentationManagerProxy.mm:
(WebKit::VideoPresentationManagerProxy::createLayerHostViewWithID):

Canonical link: https://commits.webkit.org/273707@main