[web-animations] don't normalize progress for `visibility` and `content-visibility` while animating https://bugs.webkit.org/show_bug.cgi?id=268366
Reviewed by Tim Nguyen.

For CSS properties using a "discrete" animation type, we normalize the animated value to be either
0 or 1 depending on which side of the 0.5 progress value we stand. However, the `visibility` and
`content-visibility` properties have a special type of discrete animation support that requires
to not perform this value normalization, as will the `display` property when we add animation
support for it (see bug 267762).

We introduce a new wrapper type `NonNormalizedDiscretePropertyWrapper` to do just that and use
it for the two properties listed above. This addresses the last remaining failure for the
`content-visibility` interpolation test.

* LayoutTests/imported/w3c/web-platform-tests/css/css-contain/content-visibility/content-visibility-interpolation-expected.txt:
* Source/WebCore/animation/CSSPropertyAnimation.cpp:
(WebCore::AnimationPropertyWrapperBase::normalizesProgressForDiscreteInterpolation const):
(WebCore::CSSPropertyAnimationWrapperMap::CSSPropertyAnimationWrapperMap):
(WebCore::blendStandardProperty):

Canonical link: https://commits.webkit.org/273742@main