[GStreamer][MediaStream] http/wpt/mediastream/mediastreamtrackprocessor-videoframe-timestamp.html fails https://bugs.webkit.org/show_bug.cgi?id=268215
Reviewed by Xabier Rodriguez-Calvar.

The test was failing because the mediastreamsrc was attempting to register on incoming tracks before
they were ready. This issue was not obvious until now, usually we wait the PC track events and
associate the corresponding MediaStream with a media element. But this test is using a different
approach, relying on the RTCRtpReceiver API, without waiting on track events.

* LayoutTests/platform/glib/TestExpectations:
* Source/WebCore/platform/mediastream/gstreamer/GStreamerMediaStreamSource.cpp:
(WebKitMediaStreamObserver::activeStatusChanged):
(webkitMediaStreamSrcCharacteristicsChanged):
* Source/WebCore/platform/mediastream/gstreamer/RealtimeIncomingSourceGStreamer.cpp:
(WebCore::RealtimeIncomingSourceGStreamer::hasClient):
(WebCore::RealtimeIncomingSourceGStreamer::registerClient):
* Source/WebCore/platform/mediastream/gstreamer/RealtimeIncomingSourceGStreamer.h:

Canonical link: https://commits.webkit.org/273743@main