Explore the 'serialization' issue in getAttribute('transform') https://bugs.webkit.org/show_bug.cgi?id=267401 rdar://121178816
Reviewed by NOBODY (OOPS!).

Fix the serialization of the SVG transform attribute.
After setting, for example,
    svg.createSVGTransform().setScale(-2, -4)
on an SVG element, a g.getAttribute('transform') would return.
    scale(-2 -4)
This patch fixes it so it would return:
    scale(-2, -4)
as specified.
This is normalizing and fixing some tests. But we should be adding WPT
tests for this too, because this fix doesn't break any WPT tests, which
means this is not tested in WPT.

* LayoutTests/svg/dom/SVGTransformList-basics-expected.txt:
* LayoutTests/svg/dom/SVGTransformList-basics.xhtml:
* LayoutTests/svg/dom/SVGViewSpec-expected.txt:
* LayoutTests/svg/dom/SVGViewSpec-multiple-views-expected.txt:
* LayoutTests/svg/dom/SVGViewSpec-multiple-views.html:
* LayoutTests/svg/dom/SVGViewSpec.html:
* LayoutTests/svg/dom/viewspec-parser-1-expected.txt:
* LayoutTests/svg/transforms/negative-scale-value-expected.txt:
* LayoutTests/svg/transforms/negative-scale-value.html:
* Source/WebCore/svg/SVGTransformValue.h:
(WebCore::SVGTransformValue::appendFixedPrecisionNumbers):