[WGSL] Crash when failing to type check addressSpace and similar enums https://bugs.webkit.org/show_bug.cgi?id=268308 rdar://121527219
Reviewed by Mike Wyrzykowski.

The code assumes that if we successfully unify the type of the expression with
the addressSpace enum type, then the node must be an IdentifierExpression. This
is valid most of the time, except the type is bottom, then the expression might
be anything since bottom with unifies with all types. To fix that we explicitly
check if the type is bottom.

* Source/WebGPU/WGSL/TypeCheck.cpp:
(WGSL::TypeChecker::texelFormat):
(WGSL::TypeChecker::accessMode):
(WGSL::TypeChecker::addressSpace):

Canonical link: https://commits.webkit.org/273749@main