[Cocoa] Introduce WKSLinearMediaPlayer https://bugs.webkit.org/show_bug.cgi?id=268361 rdar://121907476
Reviewed by Jer Noble.

Declared an Objective-C interface for WKSLinearMediaPlayer and a protocol for
WKSLinearMediaPlayerDelegate. The interface is implemented in Swift, where it conforms to the
Playable protocol. The properties exposed to Objective-C are bound to Publishers via key-value
observing, and Playable functions are exposed to Objective-C via the delegate. Since
@_objcImplementation classes can only contain stored properties representable in Objective-C,
Swift-only types are stored indirectly via SwiftOnlyData.

* Source/WebKit/WebKit.xcodeproj/project.pbxproj:
* Source/WebKit/WebKitSwift/LinearMediaKit/LinearMediaPlayer.swift: Added.
* Source/WebKit/WebKitSwift/LinearMediaKit/LinearMediaTypes.swift: Drive-by fix to support Public SDK builds on visionOS.
* Source/WebKit/WebKitSwift/LinearMediaKit/WKSLinearMediaPlayer.h: Added.
* Source/WebKit/WebKitSwift/WebKitSwift.h:

Canonical link: https://commits.webkit.org/273806@main