If testRunner.notifyDone is called in a different site isolated process than testRunner.waitUntilDone, layout test doesn't complete rdar://121682316 https://bugs.webkit.org/show_bug.cgi?id=268471
Reviewed by NOBODY (OOPS!).

Before site isolation both an iframe and parent frame lived in the same
process and so used the same Injected Bundle. Done notifications worked
because even if the child frame was finishing the test it was able to
access a pointer to the main frame and dump appropriately. In site
isolation a cross origin iframe lives in a different process and so does
not have access to the main frame's pointer.

This change bypasses a main frame check when a frame is wrapping up after a
call to testRunner.notifyDone if it is not in the main frame's process.
This will allow the frame to send the "Done" notification to the UI
process which can handle notifications to the other web content
processes to finish the test.

* LayoutTests/http/tests/site-isolation/notify-done-expected.txt: Added.
* LayoutTests/http/tests/site-isolation/notify-done.html: Added.
* LayoutTests/http/tests/site-isolation/resources/frame-notify-done.html: Added.
(WTR::InjectedBundlePage::dump):