Refactor WCLayerUpdateInfo for serialization https://bugs.webkit.org/show_bug.cgi?id=267968
Reviewed by Chris Dumez.

Modify `WCLayerUpdateInfo` so it's serialization can be generated. The
values to serialize in an instance are controlled by an `OptionSet`.
The generator supports this through `OptionalTupleBit` but it expects
one field per enumeration which is not the case for `WCLayerUpdateInfo`.
Group those fields together into nested structs and sort the fields by
the value of `WCLayerChange` they correspond to.

The actual serialization of `WCLayerUpdateInfo` will come in a separate
change.

* Source/WebKit/GPUProcess/graphics/wc/WCScene.cpp:
(WebKit::WCScene::update):
* Source/WebKit/WebProcess/WebPage/wc/GraphicsLayerWC.cpp:
(WebKit::GraphicsLayerWC::flushCompositingStateForThisLayerOnly):
* Source/WebKit/WebProcess/WebPage/wc/WCUpdateInfo.h:
(WebKit::WCLayerUpdateInfo::encode const):
(WebKit::WCLayerUpdateInfo::decode):
* Source/WebKit/WebProcess/WebPage/wc/WCUpdateInfo.serialization.in:

Canonical link: https://commits.webkit.org/273839@main