Explicitly invalidate grant for locked files https://bugs.webkit.org/show_bug.cgi?id=268473 rdar://121746578
Reviewed by Chris Dumez.

Explicitly invalidate grant for locked files in the Networking process. This is a speculative fix for a crash
we're seeing in the Networking process because it has taken too many assertions. The assertions are currently
being invalidated when the grant object is being destroyed. The grant object is autoreleased, and will
therefore be destroyed in an autorelease pool. The reasoning behind this change is that we should explicitly
invalidate the assertions without relying on when they are being destroyed, since that is outside of WebKit
control. This patch also adds some new logs and faults.

* Source/WebKit/NetworkProcess/cocoa/NetworkProcessCocoa.mm:
(WebKit::NetworkProcess::acquireLockedFileGrant):
(WebKit::NetworkProcess::invalidateGrant):
* Source/WebKit/Shared/AuxiliaryProcessExtensions/NetworkingProcessExtension.swift:
(Grant.invalidateGrant):
(NetworkingProcessExtension.grant(_:name:)):
(Grant.invalidate): Deleted.
* Source/WebKit/Shared/Cocoa/WKProcessExtension.h:
* Source/WebKit/Shared/Cocoa/WKProcessExtension.mm:
(-[WKGrant invalidateGrant]):

Canonical link: https://commits.webkit.org/273865@main