Navigations to data URLs should be loaded in the frame's current process with site isolation on https://bugs.webkit.org/show_bug.cgi?id=268577 rdar://122132623
Reviewed by NOBODY (OOPS!).

A WPT test was crashing with site isolation on because of issues in our process selection code
because data URLs don't have a RegistrableDomain.  To consistently choose a process and prevent
unnecessary process launches, just load a data URL in the same process as it was already using
for its web content.

* LayoutTests/platform/mac-site-isolation/TestExpectations:
* Source/WebKit/UIProcess/WebProcessPool.cpp:
(WebKit::WebProcessPool::processForNavigation):